Correcting reported conflict.
This error can't be replicated on my end so it's possible it's caused by using an earlier version of Electric Expansion, but I've changed it anyway.

Because this change isn't reflected in the ID dump I can only assume this particular block has been removed in newer versions; consequently no update to the mIDas patches has been pushed.