Massive mod support extension and second ID pass
A major effort has been made to unify mod block/item IDs to get them to use specific ranges, and especially in the case of mods with addons the addon mods have had their IDs changed as best as possible to bring them closer to the ranges their parent mod uses.

This process was not totally successful, and TiCo and Forestry still have their bizarre ID range spread.

Ultimately this process makes it considerably simpler to both add extra mods to the game and to read the NEI block/item list in full in-game. Further development is strongly encouraged to stick to this practice.